separate cable_core from cable, update dependencies and readmes
