+++ Cargo.toml
@@ -2,5 +2,6 @@
 
 members = [
     "cable",
+    "cable_core",
     "desert"
 ]
