+++ README.md
@@ -2,4 +2,10 @@
 
 Experimental [cable](https://github.com/cabal-club/cable) protocol implementation in Rust.
 
-Also included in this workspace is a vendored version of the [desert](https://crates.io/crates/desert) library by substack.
+**Status**: alpha (under active construction; expect changes).
+
+**Crates**
+
+[cable]("cable") : Cable binary payload encoding and decoding (plus post and message types)
+[cable_core]("cable_core) : Store and stream implementations for creating cable peers
+[desert]("desert") : Serialization and deserialization traits (vendored version; authored by substack)
