+++ cable/Cargo.toml
@@ -6,12 +6,6 @@ edition = "2021"
 [dependencies]
 sodiumoxide = "0.2.7"
 desert = { path = "../desert" }
-async-std = { version = "1.10.0", features = ["attributes","unstable"] }
-futures = "0.3.13"
-#length-prefixed-stream = "1.0.0"
-signature = "1.3.1"
-async-trait = "0.1.51"
 
 [dev-dependencies]
-argmap = "1.1.1"
 hex = "0.4.3"
