+++ cable_core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cable_core"
+version = "1.0.0"
+edition = "2021"
+
+[dependencies]
+async-std = { version = "1.10.0", features = ["attributes","unstable"] }
+async-trait = "0.1.51"
+cable = { path = "../cable" }
+futures = "0.3.13"
+desert = { path = "../desert" }
+#length-prefixed-stream = "1.0.0"
+signature = "1.3.1"
+sodiumoxide = "0.2.7"
+
+[dev-dependencies]
+argmap = "1.1.1"
