+++ cable_core/README.md
@@ -1,0 +1,7 @@
+# cable_core
+
+Store and stream implementations to facitilate creation of cable peers.
+
+**Status**: alpha (under active construction; expect changes).
+
+This library does not handle encoding and decoding cable binary payloads. See the [cable]("../cable") crate for those features.
