+++ cable_core/src/stream.rs
@@ -10,8 +10,9 @@ use async_std::{
     task,
     task::{Context, Poll, Waker},
 };
+use cable::{error::Error, post::Post, Hash};
 
-use crate::{error::Error, post::Post, ChannelOptions, Hash};
+use crate::ChannelOptions;
 
 /// An asynchronous stream of posts.
 pub type PostStream<'a> = Box<dyn Stream<Item = Result<Post, Error>> + Unpin + Send + 'a>;
